Making a dedicated GitPathView
