+++ src/UnityExtension/Assets/Editor/GitHub.Unity/GitHub.Unity.csproj
@@ -102,6 +102,7 @@
     <Compile Include="UI\HistoryView.cs" />
     <Compile Include="UI\IView.cs" />
     <Compile Include="UI\PublishView.cs" />
+    <Compile Include="UI\GitPathView.cs" />
     <Compile Include="UI\SettingsView.cs" />
     <Compile Include="UI\Subview.cs" />
     <Compile Include="UI\Window.cs" />
